Channels api calls implemented.

+++ RabbitMQWebAPI.Library/Models/Channel/ChannelGarbageCollection/ChannelGarbageCollectionSentinel.cs
@@ -13,11 +13,11 @@ namespace RabbitMQWebAPI.Library.Models.Channel.ChannelGarbageCollection
         public override IModel ParseDictionaryToParameters(IDictionary<String, Object> parametersDictionary)
         {
             ChannelGarbageCollection model = new ChannelGarbageCollection();
-            model.fullsweep_after = Int32.Parse(parametersDictionary["fullsweep_after"].ToString());
-            model.max_heap_size = Int32.Parse(parametersDictionary["max_heap_size"].ToString());
-            model.min_bin_vheap_size = Int32.Parse(parametersDictionary["min_bin_vheap_size"].ToString());
-            model.min_heap_size = Int32.Parse(parametersDictionary["min_heap_size"].ToString());
-            model.minor_gcs = Int32.Parse(parametersDictionary["minor_gcs"].ToString());
+            model.fullsweep_after = double.Parse(parametersDictionary["fullsweep_after"].ToString());
+            model.max_heap_size = double.Parse(parametersDictionary["max_heap_size"].ToString());
+            model.min_bin_vheap_size = double.Parse(parametersDictionary["min_bin_vheap_size"].ToString());
+            model.min_heap_size = double.Parse(parametersDictionary["min_heap_size"].ToString());
+            model.minor_gcs = double.Parse(parametersDictionary["minor_gcs"].ToString());
 
             return model;
         }
