+++ RabbitMQWebAPI.Library/Models/Exchange/ExchangeMessageStats/ExchangeMessageStatsSentinel.cs
@@ -15,40 +15,40 @@ namespace RabbitMQWebAPI.Library.Models.Exchange.ExchangeMessageStats
         public override IModel ParseDictionaryToParameters(IDictionary<String, Object> parametersDictionary)
         {
            ExchangeMessageStats parameters = new ExchangeMessageStats();
-            parameters.publish_in = Int32.Parse(parametersDictionary["publish_int"].ToString());
+            parameters.publish_in = double.Parse(parametersDictionary["publish_int"].ToString());
             parameters.publish_in_details =
-                JsonConvert.DeserializeObject<Dictionary<string, int>>(
+                JsonConvert.DeserializeObject<Dictionary<string, double>>(
                     parametersDictionary["publish_in_details"].ToString());
 
-            parameters.publish = Int32.Parse(parametersDictionary["publish"].ToString());
+            parameters.publish = double.Parse(parametersDictionary["publish"].ToString());
             parameters.publish_details =
-                JsonConvert.DeserializeObject<Dictionary<string, int>>(
+                JsonConvert.DeserializeObject<Dictionary<string, double>>(
                     parametersDictionary["publish_details"].ToString());
 
-            parameters.publish_out = Int32.Parse(parametersDictionary["publish_out"].ToString());
+            parameters.publish_out = double.Parse(parametersDictionary["publish_out"].ToString());
             parameters.publish_out_details =
-                JsonConvert.DeserializeObject<Dictionary<string, int>>(parametersDictionary["publish_out_datails"].ToString());
+                JsonConvert.DeserializeObject<Dictionary<string, double>>(parametersDictionary["publish_out_datails"].ToString());
 
-            parameters.ack = Int32.Parse(parametersDictionary["ack"].ToString());
+            parameters.ack = double.Parse(parametersDictionary["ack"].ToString());
             parameters.ack_details =
-                JsonConvert.DeserializeObject<Dictionary<string, int>>(parametersDictionary["ack_details"].ToString());
+                JsonConvert.DeserializeObject<Dictionary<string, double>>(parametersDictionary["ack_details"].ToString());
 
-            parameters.deliver_get = Int32.Parse(parametersDictionary["deliver_get"].ToString());
-            parameters.deliver_get_details = JsonConvert.DeserializeObject<Dictionary<string, int>>(parametersDictionary["deliver_get_details"].ToString());
+            parameters.deliver_get = double.Parse(parametersDictionary["deliver_get"].ToString());
+            parameters.deliver_get_details = JsonConvert.DeserializeObject<Dictionary<string, double>>(parametersDictionary["deliver_get_details"].ToString());
 
-            parameters.confirm = Int32.Parse(parametersDictionary["confirm"].ToString());
+            parameters.confirm = double.Parse(parametersDictionary["confirm"].ToString());
             parameters.confirm_details =
-                JsonConvert.DeserializeObject<Dictionary<string, int>>(
+                JsonConvert.DeserializeObject<Dictionary<string, double>>(
                     parametersDictionary["confirm_details"].ToString());
 
-            parameters.return_unroutable = Int32.Parse(parametersDictionary["return_unroutable"].ToString());
+            parameters.return_unroutable = double.Parse(parametersDictionary["return_unroutable"].ToString());
             parameters.return_unroutable_details = JsonConvert.DeserializeObject<
-                                                   Dictionary<string, int>>(
+                                                   Dictionary<string, double>>(
                                                        parametersDictionary["return_unroutable_details"].ToString());
 
-            parameters.redeliver = Int32.Parse(parametersDictionary["redeliver"].ToString());
+            parameters.redeliver = double.Parse(parametersDictionary["redeliver"].ToString());
             parameters.redeliver_details =
-                JsonConvert.DeserializeObject<Dictionary<string, int>>(
+                JsonConvert.DeserializeObject<Dictionary<string, double>>(
                     parametersDictionary["redeliver_details"].ToString());
 
             return parameters;
